Share progress bar functions between X2Importer and StarImporter
+ Also clean up some of the logic and printing for the progress bar
